Did some debugging on the tool subclasses; the code now compiles

+++ project/iteration1/src/Eraser.h
@@ -15,7 +15,7 @@ class Eraser: public Tool
     Eraser();
     virtual ~Eraser();
     float** computeMask();
-    double calculateDistance(int x, int y);
+    float calculateDistance(int x, int y);
 };
 
 #endif
